requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.76.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
torch==1.7.1
torchvision==0.2.2.post3
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
